EventBasedRuptureCalculator is not a subclass of ClassicalCalculator anymore [skip hazardlib]
